+++ packages/react-components/src/index.ts
@@ -14,5 +14,6 @@ export * from '@fluentui/react-divider';
 export * from '@fluentui/react-image';
 export * from '@fluentui/react-link';
 export * from '@fluentui/react-menu';
+export * from '@fluentui/react-popover';
 export * from '@fluentui/react-portal';
 export * from '@fluentui/react-tooltip';