+++ tools/types.ts
@@ -6,3 +6,13 @@ export interface TsConfig {
   exclude?: Array<string>;
   references?: Array<{ path: string }>;
 }
+
+export interface PackageJson {
+  private?: boolean;
+  name: string;
+  version: string;
+  scripts?: Record<string, string>;
+  dependencies?: Record<string, string>;
+  devDependencies?: Record<string, string>;
+  peerDependencies?: Record<string, string>;
+}