+++ packages/react-cards/CHANGELOG.md
@@ -1,8 +1,17 @@
 # Change Log - @fluentui/react-cards
 
-This log was last generated on Wed, 26 May 2021 07:35:43 GMT and should not be manually modified.
+This log was last generated on Thu, 03 Jun 2021 07:36:03 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## [1.0.0-beta.101](https://github.com/microsoft/fluentui/tree/@fluentui/react-cards_v1.0.0-beta.101)
+
+Thu, 03 Jun 2021 07:36:03 GMT 
+[Compare changes](https://github.com/microsoft/fluentui/compare/@fluentui/react-cards_v1.0.0-beta.100..@fluentui/react-cards_v1.0.0-beta.101)
+
+### Changes
+
+- Bump @fluentui/react-make-styles to v9.0.0-alpha.39 ([PR #18401](https://github.com/microsoft/fluentui/pull/18401) by martinhochel@microsoft.com)
 
 ## [1.0.0-beta.100](https://github.com/microsoft/fluentui/tree/@fluentui/react-cards_v1.0.0-beta.100)